requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
ament-clang-format==0.12.11
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
bond==3.0.2
build==1.3.0
builtin-interfaces==1.2.1
cachetools==6.2.4
can-msgs==2.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.7
click==8.3.1
coloredlogs==15.0.1
composition-interfaces==1.2.1
cv-bridge==3.2.1
dataclasses-json==0.6.7
demo-nodes-py==0.20.5
diagnostic-aggregator==4.0.0
diagnostic-msgs==4.2.4
diagnostic-updater==4.0.0
distro==1.9.0
domain-coordinator==0.10.0
durationpy==0.10
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
exceptiongroup==1.3.1
fastapi==0.124.4
filelock==3.20.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
generate-parameter-library-py==0.3.8
geodesy==1.0.6
geographic-msgs==1.0.6
geometry-msgs==4.2.4
google-auth==2.45.0
googleapis-common-protos==1.72.0
gps-msgs==2.0.4
greenlet==3.3.0
grid-map-msgs==2.0.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
image-geometry==3.2.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
interactive-markers==2.3.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.2.0
langchain-chroma==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.2
langchain-openai==1.1.4
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
message-filters==4.3.5
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nav-msgs==4.2.4
nav2-common==1.1.16
nav2-msgs==1.1.16
nmea-msgs==2.0.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.13.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
osrf-pycommon==2.0.2
overrides==7.7.0
packaging==25.0
pacmod3-msgs==1.0.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-qt-binding==1.1.2
PyYAML==6.0.3
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
radar-msgs==0.2.1
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resource-retriever==3.1.2
rich==14.2.0
rmw-dds-common==1.6.0
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2test==0.4.0
ros2topic==0.18.11
rosapi==2.0.0
rosapi-msgs==2.0.0
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosbridge-library==2.0.0
rosbridge-msgs==2.0.0
rosbridge-server==2.0.0
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpds-py==0.30.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-monitor==1.0.6
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9.1
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
septentrio-gnss-driver==1.4.1
shape-msgs==4.2.4
shellingham==1.5.4
six==1.17.0
smclib==3.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
sros2==0.10.5
starlette==0.50.0
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.8
theora-image-transport==2.5.2
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
topic-monitor==0.20.5
topic-tools==1.1.1
topic-tools-interfaces==1.1.1
tqdm==4.67.1
trajectory-msgs==4.2.4
turtlesim==1.4.2
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ublox-msgs==2.3.0
udp-msgs==0.0.5
unique-identifier-msgs==2.2.1
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
velodyne-msgs==2.5.1
visualization-msgs==4.2.4
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xacro==2.0.8
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
